+ GraphQL CREATE READ operations

+++ src/groups/groups.service.ts
@@ -17,6 +17,18 @@ export class GroupsService {
   ) {
   }
 
+  async getGroups(): Promise<Group[]> {
+    return await this.groupRepository.find();
+  }
+
+  async getGroup(id: string, throwError?: boolean): Promise<Group> {
+    const group = await this.groupRepository.findOne({ id });
+    if (!group && throwError) {
+      throw new NotFoundException(`Group with id ${id} not found!`);
+    }
+    return group;
+  }
+
   async createGroup(createGroupDto: CreateGroupDto): Promise<Group> {
     const { title, users } = createGroupDto;
 
@@ -28,9 +40,8 @@ export class GroupsService {
     if (users) {
       // Add group to users from list
       for (const user of users) {
-        //Check if user exist
-        const found = await this.usersService.getUser(user);
-        if (found) {
+        // Check if user exist
+        if (await this.usersService.getUser(user)) {
           // Add group to users from list
           await this.usersService.addGroupToUser(user, group.id);
           // Add user to group's users
@@ -42,55 +53,47 @@ export class GroupsService {
     return group;
   };
 
-  async getGroups(): Promise<Group[]> {
-    return await this.groupRepository.find();
-  }
-
-  async getGroup(id: string): Promise<Group> {
-    const group = await this.groupRepository.findOne({ id });
-    if (!group) {
-      throw new NotFoundException(`Group with id ${id} not found!`);
-    }
-    return group;
-  }
-
   async updateGroup(id: string, updateGroupDto: UpdateGroupDto): Promise<Group> {
 
     const { title, users } = updateGroupDto;
 
-    const groupCopy = await this.copyGroup(id);
+    const groupCopy = Object.assign(await this.getGroup(id, true));
 
     if (title) {
       groupCopy.title = title;
     }
     // If replace list of users in group
     if (users) {
-      // Add group to users from new list
       for (const user of users) {
         // Check if we have new user not in old users list
         if (groupCopy.users.indexOf(user) === -1) {
           // Check if user exist
-          const found = await this.usersService.getUser(user);
-          if (found) {
+          if (await this.usersService.getUser(user)) {
+            // Add group to user from new list
             await this.usersService.addGroupToUser(user, groupCopy.id);
+            // Add user to group users
+            groupCopy.users.push(user);
           }
         }
       }
-      // Remove group from users from old list
       for (const user of groupCopy.users) {
-        // Check if group from old list not in new list
+        // Check if user from old list not in new list
         if (users.indexOf(user) === -1) {
-          await this.usersService.deleteGroupFromUser(user, groupCopy.id);
+          if (await this.usersService.getUser(user)) {
+            // Remove group from user who not in new list
+            await this.usersService.deleteGroupFromUser(user, groupCopy.id);
+            // Remove user from group users
+            groupCopy.users = groupCopy.users.filter(item => item !== user);
+          }
         }
       }
-      groupCopy.users = users;
     }
     await groupCopy.save();
     return groupCopy;
   }
 
   async deleteGroup(id: string): Promise<void> {
-    const group = await this.getGroup(id);
+    const group = await this.getGroup(id, true);
     // Remove deleted group from all users
     for (const user of group.users) {
       await this.usersService.deleteGroupFromUser(user, group.id);
@@ -99,19 +102,15 @@ export class GroupsService {
   }
 
   async addUserToGroup(userId: string, groupId: string): Promise<void> {
-    const groupCopy = await this.copyGroup(groupId);
+    const groupCopy = Object.assign(await this.getGroup(groupId));
     groupCopy.users.push(userId);
     await groupCopy.save();
   }
 
   async deleteUserFromGroup(userId: string, groupId: string): Promise<void> {
-    const groupCopy = await this.copyGroup(groupId);
+    const groupCopy = Object.assign(await this.getGroup(groupId));
     groupCopy.users = groupCopy.users.filter(user => user !== userId);
     await groupCopy.save();
-  }
-
-  async copyGroup(groupId): Promise<Group> {
-    return Object.assign(await this.getGroup(groupId));
   }
 
   async getManyGroups(groupsIds: string[]): Promise<Group[]> {
